considlate to distributed.utils
