+++ python/gigl/distributed/dataset_factory.py
@@ -32,9 +32,9 @@ from gigl.distributed.dist_context import DistributedContext
 from gigl.distributed.dist_link_prediction_dataset import DistLinkPredictionDataset
 from gigl.distributed.dist_partitioner import DistPartitioner
 from gigl.distributed.dist_range_partitioner import DistRangePartitioner
-from gigl.distributed.utils import get_process_group_name
-from gigl.distributed.utils.serialized_graph_metadata_translator import (
+from gigl.distributed.utils import (
     convert_pb_to_serialized_graph_metadata,
+    get_process_group_name,
 )
 from gigl.src.common.types.graph_data import EdgeType
 from gigl.src.common.types.pb_wrappers.gbml_config import GbmlConfigPbWrapper
