+++ python/gigl/distributed/dist_range_partitioner.py
@@ -9,7 +9,7 @@ from graphlearn_torch.utils import convert_to_tensor
 
 from gigl.common.logger import Logger
 from gigl.distributed.dist_partitioner import DistPartitioner
-from gigl.distributed.utils.partition_book import get_ids_on_rank
+from gigl.distributed.utils import get_ids_on_rank
 from gigl.src.common.types.graph_data import EdgeType, NodeType
 from gigl.types.graph import FeaturePartitionData, GraphPartitionData, to_homogeneous
