--- python/gigl/distributed/utils/__init__.py
@@ -1,5 +1,0 @@
-from .device import get_available_device
-from .init_neighbor_loader_worker import (
-    get_process_group_name,
-    init_neighbor_loader_worker,
-)
